// src/general-settings/interfaces/general-settings.interfaces.ts
export enum ETicketSystem {
  ENABLED = 1,
  DISABLED = 2,
}

export enum ESignupPage {
  ENABLED = 1,
  DISABLED = 2,
}

export enum EEmailConfirmation {
  ENABLED = 1,
  DISABLED = 2,
}

export enum ETermsCheckbox {
  ENABLED = 1,
  DISABLED = 2,
}

export enum EResetPassword {
  ENABLED = 1,
  DISABLED = 2,
}